% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible_overlap.R
\name{feasible_overlap}
\alias{feasible_overlap}
\title{Feasibility of Ellipsoid Overlap}
\usage{
feasible_overlap(ell, width = 1, method = "QP", ...)
}
\arguments{
\item{ell}{a list of length at least 2 describing the ellipsoids.}

\item{width}{a vector the (non-negative) widths of ellipsoids in the
ellipsoid precision matrix norm.}

\item{method}{a character string indicating the method used for solving the
feasibility problem. The standard is "\code{QP}".}

\item{...}{additional arguments to be passed to the low level feasibility
method.}
}
\value{
logical.
}
\description{
Determin if the ellipsoids overlap.
}
\details{
long long description
}
\examples{
ell <- list(
  list(B = c(0,0), M = matrix(c(1,0,0,1), ncol = 2)),
  list(B = c(1,1), M = matrix(c(1,0,0,1), ncol = 2))
)

}
