% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible_point.R
\name{feasible_point}
\alias{feasible_point}
\title{Find An Overlap of Ellipsoids}
\usage{
feasible_point(ell, initial, method = "QP", ...)
}
\arguments{
\item{ell}{a list of length at least 2 describing the ellipsoids.}

\item{initial}{starting value for the overlap point.}

\item{method}{a character string indicating the method used for solving the
feasibility problem. The standard is "\code{QP}".}

\item{...}{additional arguments to be passed to the low level feasibility
method.}
}
\value{
logical.
}
\description{
Find the smallest width of ellipsoids resulting in an overlap.
}
\details{
long long description
}
\examples{
ell <- list(
  list(B = c(0,0), M = matrix(c(1,0,0,1), ncol = 2)),
  list(B = c(1,1), M = matrix(c(1,0,0,1), ncol = 2))
)

}
