% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_ellipsoid.R
\name{wrangle_ellipsoid}
\alias{wrangle_ellipsoid}
\title{Ellipsoid Wrangeler}
\usage{
wrangle_ellipsoid(e, method, ...)
}
\arguments{
\item{e}{a list of ellipsoids to be wrangled.}

\item{method}{a feasibility method determining the output of the wrangeler.}

\item{...}{additional arguments TODO !!}
}
\value{
The output of the wrangler depends on the method.
\describe{
  \item{QP, Marginal}{Returns the (c,P,r) parameterization.}
  \item{PairTransform}{Returns the (c,l,L) parameterization.}
}
}
\description{
short
}
\details{
Takes an ellipsoid parameterization and and outputs the parameterization
needed for \code{method} to run. A parameterization is simply a named list,
where each named entry is a parameter. The following parametrizations are
accepted as input. The radius parameter \code{r} is only needed when the
function \code{\link{feasible_overlap}} is to be applied.
\describe{
  \item{c, S, r}{A vector giving the center coordinates \code{C}, a matrix
    \code{S} specifying the dispersion and the radius parameter \code{r}:
    \deqn{latex}{E = \{x \in  R^n : (x-c)^T S^{-1} (x-c) \le r \}}}
  \item{c, P, r}{A vector giving the center coordinates \code{C}, a matrix
    \code{P} specifying the precision and the radius parameter \code{r}:
    \deqn{latex}{E = \{x \in  R^n : (x-c)^T P (x-c) \le r \}}}
  \item{c,l,L}{TODO !!!}
}
}
\examples{
# two dimensional unite ball
e <- list(c = c(0,0), S = matrix(c(1,0,0,1), ncol = 2), r = 1)
f <- wrangle_ellipsoid(e, method = "QP") # now (c,P,r) parameterization

}
